Add additional docs to `HttpClient`, allow some more methods to return errors.
This resolves the TODO that I still had pending, for documenting the additional
requirements for `HttpClient` implementations.

Some methods are also updated to return absl::Status types, since it's become clear
that there will be some situations in which implementers will want to be able to cut
short a request (e.g. when they receive an unexpected/unsupported header).

Lastly, one method is also marked with ABSL_MUST_USE_RESULT, to help hint
at the fact that the caller is supposed to do something with its results (i.e. it
doesn't do anything one its own). We could've marked more methods as such as well,
but those are all 'const' already which already serves as a hint on its own.

PiperOrigin-RevId: 410103028