Make fl_runner.cc record network stats in opstats once per_phase_logging is on.
This way the FederatedProtocol implementations don't need to interact with
OpStatsLogger anymore (once the flag is rolled out), and a few code paths are
simplified.

To make testing easier, we also extend MockFederatedProtocol to keep track of which
phase of the protocol has been reached, and to return different sets of network
stats and different retry windows after each stage. This way unit tests don't have
to mimic/mock the retry window/network stats behavior anymore, which simplifies
things a little bit.

PiperOrigin-RevId: 421916376