Use legacy gRPC protocol's `TaskEligibilityInfo` proto in new HTTP protocol.
Update task_assignments.proto to use the legacy gRPC protocol's
`TaskEligibilityInfo` definition, because which proto is output is actually
a contract between the task payloads/plan.proto, client code, and server
protocol. We can't just define a new proto in the new protocol, because tasks
will actually need to be able to run over either the old or new protocol. In the
longer term, it might make sense to therefore move the `TaskEligibilityInfo`
definition into plan.proto instead (better reflecting that it is a part of the
task's contract), but we leave that for a subsequent change.

PiperOrigin-RevId: 423202575