Implement URI encoding helper functions, use them in HttpFederatedProtocol.
It turns out that the URI encoding logic is actually quite simple, and also that the
standard Google API docs prescribe very specific behavior that not all URI encoding
libraries would necessarily adhere to (specifically which characters should and
shouldn't be percent-encoded). See
https://github.com/googleapis/googleapis/blob/7cde5d0/google/api/http.proto#L246.

Hence, implementing the logic ourselves ends up being pretty easy. It also helps
that we only have know how to *encode* strings, which is a little easier/less
error prone than decoding them.

Note that there are two types of URI-encoding we're going to need (for single path
segments, where '/' has to be encoded; and for multiple path segments, where '/' is
left unencoded). Methods and tests are added for both kinds, but for now
HttpFederatedProtocol only uses the first kind.

PiperOrigin-RevId: 423378286