Implement `HttpFederatedProtocol::Checkin(...)`.
This implements 'regular' checkin functionality via the `StartTaskAssignmentRequest`
API, including repeated polling of the server-returned `Operation` in case the
initial response's `Operation.done != true`.

This does not yet implement the cancellation of such `Operation`s, as prescribed in
the `StartTaskAssignment` API definition. This will happen in a subsequent change.

We extract some of the existing functionality in `HttpFederatedProtocol` into a new
`ProtocolRequestHelper` class. This allows us to unit-test some behavior more
easily and centralize `ForwardingInfo`-related logic a bit more. In a subsequent
change this class will be further extended support the cancellation of pending
`Operation` requests.

PiperOrigin-RevId: 424667224