Introduce a `CallFromNativeWrapper` utility, to be used to wrap all JNI calls from native.
The utility makes it easy to forward unexpected exceptions that occur during a call
from C++ to Java via JNI to an `UncaughtExceptionHandler` (which can be used to
turn them into a crash report or something similar). This frees the native C++ code
from having to manage and propagate exceptions. Instead, the native code can simply
assume that no exceptions should ever be thrown by a JNI call, and can CHECK-fail
if that is not the case.

PiperOrigin-RevId: 430312765