Create an `HttpURLConnection`-based `HttpClientForNative` implementation.
This implementation issues each request on a different thread (provided by an
injected ExecutorService), handling the whole request in a synchronous, blocking
fashion.

Support for request cancellation (from any thread) is implemented as well, as is
support for decompressing response body payloads before handing them to the native
layer (as per the C++ `HttpClient` contract).

PiperOrigin-RevId: 430331011