Add AppendSlices and MergeAppendedSlices custom ops
These ops are designed to be used to lower the memory usage of TF graphs which
can write out some of their outputs before processing others, so that all
outputs need not be held in memory while the others finish computing.

`AppendSlices` stores checkpoints as chunks in a file, each with a footer
indicating the start position of the chunk. This allows `MergeAppendedSlices`
to read the chunks in reverse order, scanning all into memory before writing
them back out as a new, merged checkpoint.

PiperOrigin-RevId: 444324323