Add support for `x+gzip` Content-Type headers
Files served from the server with a Content-Type header set to `x+gzip` (where x is some actual content type) signifies that the client should fetch the resource, but instead of the http engine being responsible for decoding the payload, the client will decode the response after receiving the full payload.

This is only part of the work. In a future cl the client will signal to the server via the protocol that it supports x+gzip content-types.

PiperOrigin-RevId: 461145097