Implement Federated Select example iterator that fetches and serves slices.
When a federated select query is issued by the plan, we must fetch the requested
slices by constructing a URI from the template provided by the server in the
checkin response, and then feed them to the plan one by one in the same requested
order by feeding the slice data to a file and passing the file path as to the
plan as a string tesnor. The plan will then read the slices in the same way it
reads initial checkpoints.

This change also settles on the final, more consistent collection URI for federated
select queries: "internal:/federated_select" (consistent with "internal:/opstats").

PiperOrigin-RevId: 464641202