Merge FakeEventPublisher and LoggingEventPublisher to reduce duplicate code.
This reduces the amount of semi-duplicated code between the two classes, and in turn
makes it easier to modify/extend EventPublisher without having to touch as many
files.

We keep LoggingEventPublisher's original behavior of printing to stderr for all
events, as well as FakeEventPublisher's behavior of printing only error events, by
making FakeEventPublisher have a "quiet" parameter to switch between the two
behaviors.

PiperOrigin-RevId: 469582658