Group related arguments together in NetworkStats/ExampleStats structs in various classes.
This simplifies further refactorings in the future. E.g. adding a new type of stat
or removing a deprecated one will require changing fewer lines of code/method
signatures.

This CL also makes the interfaces a little more consistent in their ordering of
arguments across their methods (specifically, making sure 'error_message' arguments
are always the first argument if they are part of a method signature).

Also makes replaces the use of 'CheckIn' with 'Checkin', so that this is more
consistent across the codebase (e.g. compared to the existing FederatedProtocol::Checkin).

Also removes some arguments that are effectively ignored anyway (such as execution_index etc.).

Also makes the existing NetworkStats arguments in PhaseLogger "const NetworkStats&"
to avoid a few copies and be consistent with our newly added usage of NetworkStats
arguments in the EventPublisher and other classes.

This should be a no-op change.

PiperOrigin-RevId: 470853862