ClockStopwatch, a thread-safe stopwatch for measuring wall clock time across threads.
This measures wall clock time, such that if the stopwatch is started and stopped
from two different threads in parallel, it will measure the period of wall clock
time during which *either* of the threads' timers was running. E.g. two threads
with parallel events of 5s in length will result in a duration of 5s; two
thread with parallel events of 3s and 5s will also result in a duration of 5s.

This will be used in subsequent changes to measure the amount of time during a
computation that we are blocked on fetching Federated Select slices vs. actually
doing computations. Multiple Federated Select slices may be fetched from different
TF threads concurrently, so we need to measure actual wall clock time and not just
the time spent on each individual request.

PiperOrigin-RevId: 470854513