Make PhaseLogger/EventPublisher report "per phase network stats".
... including during the computation phase (e.g. to reflect network usage due to
the use of Federated Select).

Per-phase network stats mean that we more clearly define that a given PhaseLogger
event's network stats reflect only the network usage during the phase that event
represents. Previously this wasn't really clear and our implementation didn't
adhere to that principle: the network stats for LogCheckinCompleted actually
included data usage that was previously already reported during
LogEligibilityEvalCheckinCompleted.

We now separate the notion of "network duration" and "overall phase duration".
The former represents the actual time spent waiting on the network (as well as
decompressing data from the network), during a given stage.  The latter
represents the total time spent for that stage. This is particularly
useful/important for Federated Select, because there we'll have network usage
during the computation stage, and it would be useful to be able to separate the
time spent waiting to download slices from the time spent waiting for the whole
computation to finish.

PiperOrigin-RevId: 471389605