Create a few abstractions for secure aggregation: SecAggRunner, SecAggProtocolDelegate and SecAggRunnerFactory.
Move the majority of the secagg logic from the GrpcFederatedProtocol to SecAggRunnerImpl so that it can be reused by HttpFederatedProtocol. gRPC specific details are packaged into the implementation of GrpcSecAggProtocolDelegate.

Provide a SecAggRunnerFactory in the GrpcFederatedProtocol constructor, remove SecAggClient.  In the test, we no longer rely on checking whether SecAggClient is nullptr. Instead, we now are able to provide a mock SecAggRunner, and verify the Run method is called.

PiperOrigin-RevId: 472778796