Wire up the ResourceCache with hybrid grpc-with-http-resources stack and the full http stack.
This is mostly plumbing with the actual integration implemented in FetchResourcesInMemory. We do our best effort to get the resource out of the cache, but otherwise default to loading from the server. If we did not fetch the resource out of the cache, but the resource has a client_cache_id, then it is best effort stored in the cache, but no failures (aside from programmer errors) with the cache causes FetchResourcesInMemory to halt.

We also had to convert some `absl::string_view`s to `std::string`s because when building for android, the toolchain's standard library does not support some string_view to string conversions.

PiperOrigin-RevId: 476971595