TFLite Micro introduced the macro TF_LITE_REPORT_ERROR as an attempt to wrap the error logging functionality to check for TF_LITE_STRIP_ERROR_STRINGS (an ifdef var to ignore calls to message logging API).
This macro was meant to be used in the TFLite code that is shared with TFLite Micro. But its is being used in several places unintentionally.

This CL cleans up the usage of TF_LITE_REPORT_ERROR from TFLite code, to limit the usage of the macro to the code that TFLite shares with TFLite Micro, as this is currently not benefiting the TFLite performance.

PiperOrigin-RevId: 486223331