Further implementation of SimpleAggregationProtocol.
This change adds parts of code responsible for the protocol and client state tracking, and handling concurrent locking, and well as parts of other methods.

There are still many gaps in the coverage which can't be filled until the protocol
is fully implemented. In particular, implementation of GetStatus method is needed to validate counters of clients and distinguish failed clients from discarded.

More test coverage will be added in future changes, including coverage of concurrent calls.

PiperOrigin-RevId: 487400491