AggVector {index, value} iterator.
This is the first step towards the transition to a different pattern of iteration of values in a tensor. Instead of iteration over slices and then nested iteration over values in each slice, this hides that all behind a top level iterator that produces {index, value} pairs.

This allows us to decouple the aggregators from the specific data structures that describe sparsity.

The next step would be to remove slices from AggVector and change AggVectorAggregator to use the new iterator.

PiperOrigin-RevId: 488799553