Add initialization error logging to fl_runner, and adds more diagnostic logs to the resource cache implementation.
Adds `LogFatalInitializationError` to PhaseLogger to log errors that halt execution when the federated compute runtime is initialized, before it does any further computation.

Adds `LogNonfatalInitializationError` to PhaseLogger to log errors that do not halt execution when the federated compute runtime is initialized, before it does any further computation.

While we debug cache initialization issues, an error in FileBackedResourceCache::Create will be treated as nonfatal. This cl adds a flag to upgrade these nonfatal errors to fatal and halt execution.

Also updates the resource cache initialization logic a bit. Previously, in the case of a read failure in ReadInternal or write failure in WriteInternal, the resource cache would call Initialize(/*reset*/ true), which would delete the manifest and reinitialize the cache, deleting all currently cached artifacts. However, Initialize() calls both ReadInternal and WriteInternal, potentially causing infinite recursion. Instead, if the manifest cannot be read, it is now deleted, and the federated compute runtime bails, leaving cached resources behind. When Initialize is called on the next run, the missing manifest will be noted, recreated, and then all of the untracked cached resources will be deleted. As such, the `reset` parameter of Initialize() has been removed.

PiperOrigin-RevId: 489945463