Switch demo aggregation to use the AggregationProtocol API.
The newly developed AggregationProtocol API supports TensorFlow-less
aggregation, which helps with both performance and server inspectability.
With this change, `plan_utils.IntermediateAggregationSession` is no longer
used; it will be removed in a follow-up change.

This change largely maintains existing behavior, but there are a few
changes to metrics and HTTP error codes.

PiperOrigin-RevId: 494266931