Client support for separate aggregation authorization and client tokens.
If the server populates StartAggregationDataUploadResponse.client_token or
StartSecureAggregationResponse.client_token, that token should be used instead
of the authorization token for all future interactions. And once the server
always populates these fields, the client should only use the authorization token
for the initial request.

Since AbortAggregation requires a client token (not an authorization token), the
client should not be calling AbortAggregation before receiving a StartAggregationDataUploadResponse or StartSecureAggregationResponse.

com_google_googletest needed to be updated to support matchers for move-only types.

PiperOrigin-RevId: 496511624