Fix race conditions in HttpSecAggSendToServerImpl's abortion logic test.
The test initially attempted to ensure that the abort request is sent, despite
the interrupted_ flag being set. However, that is only true if the request is
sent within 3ms of the clock_.SetTime() call, which means that in some cases
the request would *not* be sent and instead the InterruptibleRunner would invoke
the 'abort function' or simply return a CANCELLED error before ever invoking
the runnable. The former 'abort function' invocation would subsequently cause
a segfault, because the HTTP test helper code doesn't actually initialize the
default cancellation listener std::function<void()>, which means that invoking
it triggers a segfault.

We fix the race condition, and also create two separate tests for each scenario.
We also remove the 'delayed interruptible runner' logic from the test, since it
was complicating the test and ultimately simply replicating real behavior in a
test-only helper function, which isn't that useful.

PiperOrigin-RevId: 500230943