Move Aggregation TensorData to a single array implementation.
The original design for a sparse aggregation tensor to consist of a list of data slices turned out to be inadequate for the future requirements. We are moving to
a new design where TensorData always consists of a single array of values regardless of whether the tensor is dense to sparse. Then there will be a separate "sparsity encoding" data structure that would be used to map the tensor data indices to the dense tensor indices.

This change removes all the code specific to having multiple slices of tensor data, and therefore it also temporarily removes support for sparse aggregation tensors. We'll reintroduce sparse tensors again with a new implementation.

PiperOrigin-RevId: 509576360