Demo server support for TaskAssignmentMode and PopulationEligibilitySpec.
The TaskAssignmentMode is an optional FederatedDataSource parameter that can be
used to instruct the client to use a task assignment method other than
StartTaskAssignment. (Note that client support is not yet implemented.) The
TaskAssignmentMode is plumbed through the server to the EligibilityEvalTasks
service, where it's exposed via a PopulationEligibilitySpec message to clients
that support it.

The TaskAssignmentMode is also used to partition tasks in the TaskAssignments
service so that each method only returns tasks of the corresponding type.

PiperOrigin-RevId: 516948681