Fix GCC build breakage caused by updating to TensorFlow 2.11.0.
TensorFlow 2.11.0 added `--std=c++17` to the flags returned by
`tf.sysconfig.get_compile_flags()`; GCC throws an error if this flag is used
when compiling C code, so the previous approach of adding all TF compile flags
to `//command_line_option:copt` is wrong. The flags don't change often, so it's
feasible to maintain an explicit list of flags that should be C++-only.

This is expected to fix the issue reported in
https://github.com/google/federated-compute/issues/1.

PiperOrigin-RevId: 518604361