Create InputTensorList type to be used as an input to operations that may take multiple tensors.
The type contains const pointers to tensors and has an iterator meeting the constraints of a forward_iterator (https://en.cppreference.com/w/cpp/iterator/forward_iterator)

The type contains either an inlined array on the stack, or an allocated vector on the heap, depending on the number of tensors it needs to store. This should make it efficient for small numbers of tensors.

PiperOrigin-RevId: 524910229