Create a CompositeKeyCombiner class operating on groups of tensors of a predefined list of datatypes to combine indices for which the same combination of elements occurs, or in other words, indices containing the same composite key.
This class contains two methods: Accumulate and GetOutputKeys, which can each
be called multiple times.

Accumulate takes in an InputTensorList of tensors of the same shape, and
returns a Tensor of the same shape containing ordinals to represent the
composite key that exists at each index. Composite keys are stored
across calls to Accumulate, so if the same composite key is ever encountered
in two different indices, whether in the same or a different call to
Accumulate, the same ordinal will be returned in both these indices.

GetOutputKeys returns the composite keys that have been seen in all previous
calls to Accumulate, represented by a vector of Tensors. If the ordinal
returned by Accumulate for that composite key was i, the composite key will
be found at position i in the output vector.

PiperOrigin-RevId: 525763609