Add PARTIAL to the ContributionResult enum, and add successfully contributed task names to the FLRunnerResult.
With the introduction of the new PARTIAL enum, the ContributionResult enum is used as following:
PARTIAL is used when at least one of the tasks (including the ones from the multiple task assignments and single task assignment) successfully contributed result.
FAIL is used when all of the tasks failed to contribute result.
SUCCESS is used when all of the tasks successfully contributed results.

After this change, we no longer return early when single task assignment is rejected or failed. Instead, we tracked which tasks are successful via a vector of successfully contributed task names and also keep track the total number of expected tasks to run and update.  At the end of `RunFederatedComputation`, we compare the number of successful tasks and the expected tasks, and use the rule above to determine which ContributionResult to return.

Note for single task assignment (check-in), we always increment the number of expected tasks by 1 because when we call `IssueCheckin` we expect the server to assign 1 task.  If the server rejects the client, we will have zero successful task, and hence we'll return FAIL ContributionResult. Even when the eligibility eval is enabled, and all of the tasks inside the TaskEligibilityInfo have zero weights (which we technically know the server will reject this client), we still increment the expected task number to keep the current client behavior.

PiperOrigin-RevId: 532858719