SecAgg: Parallel aggregation of masked inputs (with no work on the main thread).
This is an experimental optimization of the SecAgg protocol where masked inputs messages from the clients are promptly queued, then aggregated in background by multiple parallel worker threads.

Each worker thread does the following:
1) Takes ownership of batch of inputs accumulated so far in the input queue.
2) Works asynchronously to bit-unpack all masked inputs in the batch and calculate a partial sum of masked inputs.
3) Gets an exclusive access to the total sum of masked inputs for a short time and adds the partial sum to it.

The step #2 is where the bulk of the work is done, and that can be done in parallel by multiple threads.
The step #3 can be done by only one parallel thread at a time, but that still doesn't block the main thread where new messages with masked inputs arrive.

PiperOrigin-RevId: 535340416