Add multiple task assignments events to OperationalStats, and also add PhaseStats.
In the past, we keep all the events from a single run as a list inside an OperationalStats message, and task name is a top level field for OperationalStats.  It works because we generally only run a single computation per run (we ignore the task name for eligibility eval computation)
With this change, all events from a single run will be contained inside an OperationalStats message.  Instead of having all the events in the top level "events" field, the events for each phase will be organized into separate PhaseStats.  Each PhaseStats could be associated with a different task name, and hence we can support running multiple computations in a single run, and keep the task name tracking correct.

Currently, the following Phases are supported:
1. ELIGIBILITY_EVAL_CHECKIN
2. ELIGIBILITY_COMPUTATION
3. MULTIPLE_TASK_ASSIGNMENTS
4. CHECKIN
5. COMPUTATION
6. UPLOAD

PiperOrigin-RevId: 547954886