Add information about parsing client inputs to the TFF worker configuration.
The TFF worker needs to know how to parse the client data blob into a format that can be understood by the TFF computation encoded as serialized_client_work_computation.

Use a oneof since we might want to expand to other client input types in the future. For example, eventually we will want to support the lightweight client wire format which doesn't rely on TensorFlow. And we may also want to consider formats that are produced by on-device computations other than SQL query execution.

Alternatively to adding configuration that describes the client input format, we could require that the client_work TFF computation accepts the client input data blob as a string. However, the client_work TFF computation may contain proprietary information so for transparency purposes it is better to do more work in the open source TEE implementation rather than in the TFF computation. Additionally, in the future it may be easier to use existing libraries to parse client inputs rather than adding an implementation in TFF (there are existing C++ libraries for parsing the lightweight client wire format, for example.)

PiperOrigin-RevId: 559900811