Client implementation of TfCustomPolicy support for native eligibility eval
Extends opstats to include a flag whether the client's native eligibility eval stack supports interleaved eval of fully native policies (sampling without replacement, data availability) and TensorFlow custom policies. This flag is readable from TensorFlow.

If the client does support this, and the given eligibility eval task supports it, the client will eval the entire eligibility eval task with the native eligibility stack. Fully native policies will be evaluated in C++, and TfCustomPolicies will be evaluated in TensorFlow. The client will pass an iterator to the task containing a single example with the policy name to be evaluated, and the set of task names that policy applies to. The EET will then evaluate that single policy against the given task names and whatever other client data it may use.

If the client does not support this (ie the flag is off) or the EET is too old (does not support per-policy eval, which will be indicated by having a TfCustomPolicy min version of INT_MAX), the client will fall back to evaluating the EET with the legacy eligibility eval stack.

PiperOrigin-RevId: 591843486