Fix issue where all clients reporting only empty tensors causes the wrong number of tensors to be output.
If the GroupByAggregator outputs zero tensors, which was previously the behavior when all client tensors are empty, then an InternalError will be triggered in the SimpleAggregationProtocol which causes the round to fail. However, this is not the correct behavior, as all clients uploading empty tensors is considered a valid state which should produce output with all tensors empty.

PiperOrigin-RevId: 595226385