Opt out of tasks that use eligibility policies that are unimplemented or fail to evaluate, instead of halting and dropping out.
Failing to evaluate an eligibility policy, either due to an error in the policy evaluation or due to not having a native implementation of that policy, will opt the device out of tasks using that policy instead of halting execution. This will allow legacy clients to still participate in tasks that only use policies supported by the legacy clients' runtime version, instead of halting during eligibility eval and dropping out entirely.

PiperOrigin-RevId: 599758452