Create an alternative to TfLiteWrapper, which invokes TFLite APIs only from a single thread.
The way we currently invoke TFLite could be causing some thread safety issues, because
the TFLite APIs (and the `Interpreter` instance in particular) is not documented to be
thread safe as a whole. Only certain methods are explicitly documented as thread safe.

The existing code path is kept (albeit behind some no-op refactoring to share code
with the new code path), and the new code path is flagged to allow us to test the
impact with a gradual rollout.

PiperOrigin-RevId: 599929423