Log the current index of the min sep policy as a single value instead of a map.
The MinimumSeparationPolicy now is only used for enabling round-based sampling-without-replacement. The index (i.e., the algorithmic round number) should be the same no matter what the value of the minimum separation is. Logging the index as a single value will simplify the policy implementation and decouple the eligibility determination with the policy name setting.

PiperOrigin-RevId: 601007650