Check for tstring token instances with unexpected nullptr data().
We see rare occurrences where data() returns a nullptr unexpectedly, causing the
implicit std::string conversion operator to fail.

It seems this shouldn't be allowed to happen, but nevertheless does at fairly
low frequencies. Instead of crashing in that case we'll just return an error which
may help us better diagnose the situations in which this unexpected state occurs.

PiperOrigin-RevId: 605385799