Initial implementation of the ConfidentialAggregations protocol (encrypting the payload).
This change implements most of the clients' part in the ConfidentialAggregations
protocol, without implementing any of the validations the client will eventually
need to perform. It allows clients to receive a public key from the server,
encrypt its payload with that key, and upload it to the server.

It does not yet, however, implement any of the following validations:
- checking whether the public key provided by the server is signed by the
  `oak.attestation.v1.Evidence` application-layer signing key.
- checking whether the attestation chain in `oak.attestation.v1.Evidence`
  is valid.

That functionality will follow in subsequent changes.

PiperOrigin-RevId: 610913377