Use an `AttestationVerifier` instance in `HttpFederatedProtocol`.
This instance is created by a new factory method on `SimpleTaskEnvironment`, and
is only actually used if the `enable_confidential_aggregations` flag
is turned on, in which case the `HttpFederatedProtocol` implementation will use this
verifier to verify the server-provided public key and data access policy during the
`ConfidentialAggregations` protocol.

We use an `AlwaysFailingAttestationVerifier` as the default implementation in the
`SimpleTaskEnvironment` base class, effectively disabling the confidential
aggregation protocol implementation unless a different implementation is provided by
the `SimpleTaskEnvironment` implementor.

PiperOrigin-RevId: 614830373