Define a new format dedicated to clients' payloads in the ConfidentialAggregations protocol.
This updates `http_federated_protocol.cc` to stop using the `Record` proto to hold clients'
ciphertext + keys + additional metadata, and switches to a new (somewhat bespoke) format
dedicated solely to serving the client-to-server upload function.

Two considerations were taken into account:
- In order to support efficiently handling incoming uploads in a streaming fashion, the
  ciphertext is transmitted alongside (rather than within) the header proto. This
  ensures that we can, if we so choose, parse and read all of the header data before
  we've even received/read all of the ciphertext data.
- To keep the option open of migrating to another header format in the future (e.g.
  a non-proto-based one), we prepend a known byte sequence to the start of every encoded
  payload.

This change also implements compression of the unencrypted data before encryption, which
should reduce the ciphertext size a bit. Note that in an earlier change we already disabled
compression of the outer payload during uploads for the ConfidentialAggregations protocol.

PiperOrigin-RevId: 619289011