Store composite key representations in an alignment-safe structure.
CompositeKeyCombiner represents keys of every type as a uint64_t. It guarantees that the representation of all types it stores is no larger than `sizeof(uint64_t)`.

However, prior to this change composite keys were stored in a string of allocated size `sizeof(uint64_t) * tensors.size()`. Since strings are made of 1 byte characters, the string may or may not be allocated at an address that is aligned for storing `uint64_t`.

Depending on the compiler, this may cause pointer alignment issues when writing values of size greater than `sizeof(char)` to the address. Since this is a compiler-dependent problem, it was only discovered when compiling the code in a new codebase.

This change fixes the problem by using a FixedArray<uint64_t> rather than a string as the data buffer for the uint64_t values. This structure will be allocated on an address that is properly aligned to store uint64_t.

This codebase previously avoided absl dependencies, but we no longer expect to use it in the Oak Restricted Kernel, only Oak Containers, so absl dependencies are now acceptable.

Also move to using absl::node_hash_map (which has pointer stability, like std::unordered_map) since it has the built-in ability to hash absl::FixedArray, and is generally recommended over std::unordered_map.

PiperOrigin-RevId: 619563778