Fix bug with relying on argument order when creating an output tensor in OnDimGroupingAggregator.
The C++ standard does not specify in what order arguments will be evaluated: see https://stackoverflow.com/questions/2934904/what-is-the-order-of-evaluation-for-function-arguments. So depending on the compiler, it's possible that the data_vector_ will be moved from within the move constructor for the `data` parameter before the second argument is evaluated to get the vector size. This could lead to accessing invalid memory when calling the size() method on the data vector.

Instead, the code should calculate the vector size ahead of time.

PiperOrigin-RevId: 619646539