Avoid invoking the `tstring tstring(const tstring&)` copy constructor before comparing `token.data()` to `nullptr`.
The previous fix in this location attempted to avoid a nullptr SIGSEGV when invoking the `std::string operator()`
conversion function, but of course the attempted fix caused us to still invoke a similar `tstring tstring(const tstring&)`
copy constructor.

This change ensures we avoid that copy constructor, and hopefully gets us the originally intended effect.

PiperOrigin-RevId: 622932565