Add some initial documentation about attestation of server-side TEE-hosted applications.
This describes how the attestation verification library is used as part of the
client-side implementation of the `ConfidentialAggregations` protocol. We also
describe how the attestation verification performed by the code in this repository
relates to the TEE-hosted applications maintained in the Confidential Federated
Compute repository [1].

Note that these docs are expected to be expanded on in the coming
days/weeks.

[1] https://github.com/google-parfait/confidential-federated-compute

PiperOrigin-RevId: 625851984