Add a utility for extracting serialized attestation records from a log stream.
This utility can process a log stream (e.g. as provided by `adb logcat`) and
extract deserialized log records from the stream, writing each record to its own
file (as a serialized `AttestationVerificationRecord` proto). These files can
subsequently be inspected further.

Also fixes a bug in the code that logs the attestation records in chunks of 200
chars at a time, when run on Android. Instead of using the "%.*s" format string
we were previously using the incorrect "%*s" format string. The former ensures
that only a certain number of characters (200 in our case) of the provided
C string are printed, while the latter interprets the length argument as
a padding specification (which is subsequently ignored, since no padding
character is specified in the format string). This caused us to print much more
than 200 characters or the data at a time, resulting in log lines containing
partially overlapping data in each line (each line up to 1023 characters long,
when tested on my device).

The utility for extracting records from a log stream works around this issue by
heuristically detecting such overlong lines and truncating them to 200 chars.

PiperOrigin-RevId: 627907465