Support multiple src and dst node ids per transform
With the Ledger API, a transform worker can consume multiple different
sources by matching multiple Transforms in the access policy. Since the
KeyManagementSerive API pre-authorizes transform workers, this means
that the AuthorizeConfidentialTransformResponse needs to support
multiple sources (already supported) and multiple destinations (fixed
by this change). It's also necessary to re-add destination node ids to
the access policy so that the KeyManagementService can derive a
different encryption key for each.

Matching multiple transforms with identical ReferenceValues is
wasteful, so this change also allows multiple source and destination
node ids to be specified on a single Transform. In addition to making
access policies more concise, this makes it easier to represent a
mismatched number for sources and destination node ids.

PiperOrigin-RevId: 745341228