Implement `get_model` in `ProgramInputProvider`.
`ProgramInputProvider` receives a map `model_id_to_zip_file` which contains the mapping from a model id to a zip file containing the model. When the program asks to load a model with a given `model_id`, it will unzip the model file to a directory named with the model id and load the model using `tff.learning.models.load_functional_model`.

PiperOrigin-RevId: 780759889