Modify privacy ID utils to better suit usage:
- Update SplitResultsByPrivacyId to no longer add the privacy ID to the ExampleQueryResult.

Because ExampleQueryPlanEngine uses the ExampleQuery output vector specs to generate aggregation tensors, it doesn't generate an aggregation tensor for the privacy ID column since that column isn't in the output vector specs. Since the PerPrivacyIdResult has the privacy ID as a separate field already and the ExampleQueryPlanEngine will have to generate the tensor for the privacy ID column separately anyway, there's no need to put it in the ExampleQueryResult.

- Allow `PrivacyIdConfig` to not specify a windowing schedule.

If `PrivacyIdConfig` does not include a `windowing_schedule`, the `source_id` is used directly as the privacy ID, effectively grouping all events under a single privacy ID.

- Add validation that privacy ID window schedule has a positive window size.

Privacy IDs are user IDs that optionally rotate. They are stored in the encrypted portion of confidential aggregation uploads, which means they're only access within the data processing TEEs described in the associated data access policy. Currently they are computed using a source ID, a per-task persistent random ID.

These privacy IDs will make it possible to achieve user-level DP guarantees across multiple data uploads from a single device. This fits into the broader project to make the scope at which an ε/δ-differential privacy guarantee applies verifiable.

Partition keys are coarse-grained proxies for privacy IDs - many privacy IDs map to the same partition key. We limit the granularity of partition keys. They are stored in the clear and will be used by the untrusted server to efficiently process uploads associated with the same privacy ID in a grouped manner.

PiperOrigin-RevId: 797049908