Add utilities for SignedPayload verification.
The SignedPayload message holds zero or more signatures of the same payload,
like the COSE_Sign structure. Signatures may be directly signed using a key
known to the verifier or may contain a chain of signatures rooted in a known
key. Signatures may also be included in a transparency log.

These utilities will be used both for verifying HPKE upload encryption keys
and access policy endorsements.

PiperOrigin-RevId: 814040173