Add AttestationTransparencyVerifier.
This AttestationVerifier ensures that the access policy and KMS evidence
have been uploaded to a transparency log, allowing offline verification
of the TEE evidence to be performed. This verifier doesn't check the
evidence itself, but it does ensure that the necessary information to
verify the TEE evidence has been irrevocably added to a public log
(assuming the settings require transparency log entries, which they
should for all use-cases that require transparency).

Avoiding a dependency on the Oak Attestation Verification library also
avoids transitive dependencies on Rust, a JSON parser, a CBOR parser,
and RustCrypto, reducing binary size and build system complexity. It also
avoids a dependency on the AMD SEV-SNP attestation format, which is not
guaranteed to be backwards-compatible.

PiperOrigin-RevId: 814315599